#8 En passant works (i think)
ugly solution of en passant. As usual i will probably refactor later just want it to work. So now we let board have a memory of en_passant as a position. each time we move a piece two steps forward we will set en_passant to to_pos-direction position. wordy explained, but that works as we only will have one en_passant at the time. Then each move we check is en_passant, which checks for the piece moved is pawn, and that the move is capture movement. if it is en passant we will remove the piece that is en passanted, and keep it in memory look for checks, if check place everythin back as it was.